Following Pluralsight course
I've been learning data structures and generics. The generics course on Pluralsight had a short course where it went through a bunch of examples of generic data structures.

So I implemented them. Here's that commit with some comments which are essentially notes on what I understood the datastructures to be useful for.